% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adme_filter.R
\name{admet_filter}
\alias{admet_filter}
\title{ADMET 过滤}
\usage{
admet_filter(compound_df)
}
\arguments{
\item{compound_df}{包含化合物名称和 InChIKey 的数据框。}
}
\value{
返回一个数据框，经过 ADMET 过滤后的结果。
}
\description{
这个函数接受化合物数据框，并应用 ADMET 过滤。
}
\examples{
compound_df <- data.frame(
  NameEN = c("beta-Alanine", "Pyrocatechol", "2,3-Dihydroxybenzoic acid (Pyrocatechuic acid)", "Glycine", "Hydroquinone", "Lysine", "Poncirin", "Vanillin", "Xanthine"),
  inchikey = c("UCMIRNVEIXFBKS-UHFFFAOYSA-N", "YCIMNLLNPGFGHC-UHFFFAOYSA-N", "GLDQAMYCGOIJDV-UHFFFAOYSA-N", "DHMQDGOQFOQNFH-UHFFFAOYSA-N", "QIGBRXMKCJKVMJ-UHFFFAOYSA-N", "KDXKERNSBIXSRK-YFKPBYRVSA-N", "NLAWPKPYBMEWIR-YZOWQMJMSA-N", "MWOOGOJBHIARFG-UHFFFAOYSA-N", "LRFVTYWOQMYALW-UHFFFAOYSA-N")
)
result <- admet_filter(compound_df)
}
