% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autodock_accute_dock.R
\name{autodock_accute_dock}
\alias{autodock_accute_dock}
\title{精准 Autodock 分子对接函数}
\usage{
autodock_accute_dock(protein_path, ligand_path, output_dir, tool_id)
}
\arguments{
\item{protein_path}{蛋白质文件的路径。}

\item{ligand_path}{配体文件的路径。}

\item{output_dir}{输出文件夹的路径，默认为 "results/dock"。}

\item{tool_id}{切换快速对接和精准对接的ID，默认为精准对接id"fdd780e7acf24f37"；快速对接id为"aee83f8c6aa459ef"。}
}
\value{
返回一个数据框，包含精准对接的分数和排名。
}
\description{
这个函数用于进行精准的 Autodock 分子对接分析，调用了 \code{dock_autodock} 函数进行对接，并自动处理对接结果（如提取评分），最后保存为 Excel 文件。
}
\details{
该函数首先调用 \code{dock_autodock} 进行精准对接，之后提取对接结果文件中的分数，并保存为 Excel 文件。
}
\examples{
/dontrun{
accute_result <- autodock_accute_dock(
  protein_path = "/path/to/protein.pdb",
  ligand_path = "/path/to/ligand.sdf",
  output_dir = "/path/to/output",
  tool_id = "fdd780e7acf24f37"
)
}
}
\keyword{accurate,}
\keyword{autodock,}
\keyword{dock,}
\keyword{excel}
\keyword{results,}
