% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autodock_fast_dock.R
\name{autodock_fast_dock}
\alias{autodock_fast_dock}
\title{Autodock 快速分子对接与处理结果函数}
\usage{
autodock_fast_dock(protein_path, ligand_path, output_dir, tool_id)
}
\arguments{
\item{protein_path}{蛋白质文件的路径。}

\item{ligand_path}{配体文件的路径。}

\item{output_dir}{输出文件夹的路径，默认为 "results/dock"。}

\item{tool_id}{切换快速对接和精准对接的ID，默认为快速对接id"aee83f8c6aa459ef"；精准对接id为"fdd780e7acf24f37"。}
}
\value{
返回一个数据框，包含对接的分数和排名。
}
\description{
这个函数进行 Autodock 分子对接，随后自动处理对接结果（如文件拆分、提取分数等），并输出到指定的 Excel 文件。
}
\details{
该函数首先调用 Autodock 分子对接工具进行对接，之后自动处理生成的对接文件（包括拆分和提取得分），并将结果保存为 Excel 文件。
}
\examples{
/dontrun{
result <- autodock_fast_dock(
  protein_path = "/path/to/protein.pdb",
  ligand_path = "/path/to/ligand.sdf",
  output_dir = "/path/to/output",
  tool_id = "aee83f8c6aa459ef"
)
}
}
\keyword{autodock,}
\keyword{biorange,}
\keyword{dock,}
\keyword{excel}
\keyword{results,}
