% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dock_diffdock.R
\name{dock_diffdock}
\alias{dock_diffdock}
\title{Diffdock 分子对接函数}
\usage{
dock_diffdock(protein_path, ligand_path, samples_per_complex, output_dir)
}
\arguments{
\item{protein_path}{蛋白质文件的路径。}

\item{ligand_path}{配体文件的路径。}

\item{samples_per_complex}{每个复合物的样本数量，默认为 10。}

\item{output_dir}{输出文件夹的路径，默认为 "results/dock"。}
}
\value{
无返回值，但会在指定的目录下生成 Diffdock 分子对接的结果。
}
\description{
这个函数用于进行 Diffdock 分子对接分析，调用了 biorange 包的 diffdock_run 函数。
}
\details{
这个函数主要用于进行 Diffdock 的分子对接分析。
}
\examples{
/dontrun{
dock_diffdock(
  protein_path = "/home/liuyan/projects/package/bioranger/inst/example_data/6w70.pdb",
  ligand_path = "/home/liuyan/projects/package/bioranger/inst/example_data/6w70_ligand.sdf"
)
}
}
\keyword{biorange}
\keyword{diffdock,}
\keyword{dock,}
