% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dock_clean.R
\name{dock_fix_protein}
\alias{dock_fix_protein}
\title{修复蛋白质结构}
\usage{
dock_fix_protein(input_path, output_path)
}
\arguments{
\item{input_path}{字符串，表示输入蛋白质 PDB 文件的路径。}

\item{output_path}{字符串，表示修复后 PDB 文件的保存路径。}
}
\value{
返回一个表示修复结果的对象。
}
\description{
使用 \code{biorange} 库修复蛋白质的 PDB 文件。
}
\details{
此函数使用 \code{biorange} 库的 \code{dock_fix_protein} 模块处理输入的 PDB 文件。
}
\examples{
\dontrun{
result <- dock_fix_protein("data/AF-P31749-F1-model_v4.pdb", "./results/dock/fix.pdb")
}
}
\keyword{PDB文件}
\keyword{蛋白质修复,}
