% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dock_plot3d.R
\name{dock_plot3d}
\alias{dock_plot3d}
\title{3D 分子对接可视化函数}
\usage{
dock_plot3d(pdb_file, sdf_file, output_dir, width, height)
}
\arguments{
\item{pdb_file}{PDB 文件的路径。}

\item{sdf_file}{SDF 文件的路径。}

\item{output_dir}{输出文件夹的路径，默认为 "results/dock/"。}

\item{width}{输出图像的宽度，默认为 1200。}

\item{height}{输出图像的高度，默认为 900。}
}
\value{
返回一个 3D 视图对象，并在指定的目录下生成一个 HTML 文件。
}
\description{
这个函数用于进行 3D 分子对接可视化，调用了 biorange 包的 plot3d 函数。
}
\details{
这个函数主要用于生成 3D 的分子对接可视化图像。
}
\examples{
dock_plot3d(
  pdb_file = "/home/liuyan/projects/package/bioranger/inst/example_data/6w70.pdb",
  sdf_file = "/home/liuyan/projects/package/bioranger/inst/example_data/6w70_ligand.sdf"
)
}
\keyword{3d,}
\keyword{biorange}
\keyword{dock,}
\keyword{plot,}
