% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dock_smina_score.R
\name{dock_smina_score}
\alias{dock_smina_score}
\title{Smina 结合能评分函数}
\usage{
dock_smina_score(receptor_file, ligand_file)
}
\arguments{
\item{receptor_file}{受体文件的路径。}

\item{ligand_file}{配体文件的路径。}
}
\value{
返回一个包含 Smina 结合能评分的数据框。
}
\description{
这个函数用于计算 Smina 结合能评分，调用了 biorange 包的 smina_score 函数。
}
\details{
这个函数主要用于计算 Smina 的结合能评分。
}
\examples{
dock_smina_score(
  receptor_file = "/home/liuyan/projects/package/bioranger/inst/example_data/6w70.pdb",
  ligand_file = "/home/liuyan/projects/package/bioranger/inst/example_data/6w70_ligand.sdf"
)
}
\keyword{biorange}
\keyword{dock,}
\keyword{score,}
\keyword{smina,}
