% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dock_split.R
\name{dock_split_sdf}
\alias{dock_split_sdf}
\title{生成分子对接结果}
\usage{
dock_split_sdf(ligand_path)
}
\arguments{
\item{ligand_path}{字符串，表示配体SDF文件的路径。}
}
\value{
返回一个表示对接结果的对象。
}
\description{
此函数用于从SDF文件生成分子对接结果。
}
\details{
此函数使用biorange库从提供的配体文件生成对接结果。
}
\examples{
\dontrun{
result <- dock_split_sdf("/home/liuyan/projects/package/bioranger/results/dock/ligand1_docked__d49efd3e-6a2e-4a4f-aeac-1b7818c479eb", "results/dock/")
}
}
\keyword{SDF文件}
\keyword{分子对接,}
