% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockplot_2d.R
\name{dockplot_2d}
\alias{dockplot_2d}
\title{生成2D对接图}
\usage{
dockplot_2d(protein_path, ligand_path, output_path)
}
\arguments{
\item{protein_path}{字符串，表示蛋白质PDB文件的路径。}

\item{ligand_path}{字符串，表示配体SDF文件的路径。}

\item{output_path}{字符串，表示保存输出HTML文件的路径。}
}
\value{
返回一个表示2D对接图的对象。
}
\description{
此函数用于生成并保存2D分子对接图。
}
\details{
此函数使用biorange库从提供的蛋白质和配体文件生成2D图。
}
\examples{
\dontrun{
protein_file <- "path/to/rec.pdb"
ligand_file <- "path/to/vina_output.sdf"
dockplot_2d(protein_file, ligand_file, "results/dock/saving.html")
}
}
\keyword{2D图}
\keyword{分子对接,}
\keyword{可视化,}
