% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dock_plot3ds.R
\name{dockplot_3d}
\alias{dockplot_3d}
\title{Generate 3D Docking Plot}
\usage{
generate_dock_plot_3d(protein_path, ligand_path, output_path)
}
\arguments{
\item{protein_path}{A string representing the path to the protein PDB file.}

\item{ligand_path}{A string representing the path to the ligand SDF file.}

\item{output_path}{A string representing the path to save the output HTML file.}
}
\value{
An object representing the 3D docking plot.
}
\description{
This function generates and saves a 3D docking plot.
}
\details{
This function utilizes the biorange library to generate a 3D plot
from the provided protein and ligand files.
}
\examples{
protein_file <- "path/to/rec.pdb"
ligand_file <- "path/to/vina_output.sdf"
dockplot_3d(protein_file, ligand_file, "results/dock/dd.html")
}
\keyword{3D}
\keyword{docking,}
\keyword{plot}
\keyword{visualization,}
