% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrich_gokegg.R
\name{enrich_gokegg}
\alias{enrich_gokegg}
\title{GO 和 KEGG 富集分析函数}
\usage{
enrich_gokegg(gene_list, analysis_type, organism, kegg_database, go_databases, p_value_cutoff)
}
\arguments{
\item{gene_list}{一个包含基因列表的向量。}

\item{analysis_type}{分析类型，可以是 "all"、"go" 或 "kegg"，默认为 "all"。}

\item{organism}{物种名称，默认为 "human"。}

\item{kegg_database}{KEGG 数据库的名称，默认为 "KEGG_2021_Human"。}

\item{go_databases}{GO 数据库的列表，可以为 NULL。}

\item{p_value_cutoff}{P 值阈值，默认为 0.05。}
}
\value{
返回一个包含富集分析结果的数据框。
}
\description{
这个函数用于进行 GO 和 KEGG 富集分析，调用了 biorange 包的 enrich_gokegg 函数。
}
\details{
这个函数主要用于进行 GO 和 KEGG 的富集分析，并对结果进行一些后处理，以便于后续的分析。
}
\examples{
# KEGG
kegg_res <- enrich_gokegg(
  gene_list = gene_list_399,
  analysis_type = "kegg",
  p_value_cutoff = 0.05
)
head(kegg_res)
# GO
go_res <- enrich_gokegg(
  gene_list = gene_list_399,
  analysis_type = "go",
  p_value_cutoff = 0.05
)
head(go_res)
}
\keyword{biorange}
\keyword{enrichment,}
\keyword{go,}
\keyword{kegg,}
