% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_table.R
\name{format_table}
\alias{format_table}
\title{格式化表格数据}
\usage{
format_table(table_path, annotation_path)
}
\arguments{
\item{table_path}{需要格式化表格的YAML文件路径}

\item{annotation_path}{表格注释信息的YAML文件路径}
}
\value{
格式化后的表格数据
}
\description{
此函数使用指定的YAML文件格式化表格数据。
}
\details{
该函数调用biorange模块中的table_format函数，
使用提供的表格和注释路径来格式化数据。
}
\examples{
\dontrun{
formatted_table <- format_table(
    "/home/liuyan/projects/package/biorang_bak/表格路径.yaml",
    "/home/liuyan/projects/package/biorang_bak/中药入血注释.yaml"
)
print(formatted_table)
}
}
