% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{gene_list_399}
\alias{gene_list_399}
\title{339个基因名称列表}
\format{
一个字符向量，包含339个基因名称。
}
\source{
\link{数据来源，如果有的话}
}
\usage{
data(gene_list_399)
}
\description{
该数据集包含一个由339个基因名称组成的列表。
}
\examples{
data(gene_list_399)
head(gene_list_399)
}
\keyword{datasets}
