% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_target_sequence.R
\name{generate_target_sequence}
\alias{generate_target_sequence}
\title{获取化合物序列信息}
\usage{
generate_target_sequence(compound_sum, target_sum)
}
\arguments{
\item{target_sum}{一个整数，表示目标的总和。}
}
\value{
返回一个包含化合物和目标序列的列表。
}
\description{
该函数生成化合物和目标序列。
}
\examples{
compound_sum <- 8
target_sum <- 144
target_sequence <- generate_target_sequence(compound_sum, target_sum)
print("target Sequence:", target_sequence)
}
