% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_generate_type.R
\name{generate_target_type}
\alias{generate_target_type}
\title{生成靶点类型}
\usage{
generate_target_type(kegg_df, target_df)
}
\arguments{
\item{kegg_df}{富集结果 数据的数据框。}

\item{target_df}{靶点数据的数据框。}
}
\value{
返回一个列表，包含两个数据框：gf1 和 gf2。
}
\description{
这个函数接受 富集结果 和靶点数据框，并生成靶点类型。
}
\examples{
/dontrun{
kegg_df <- read.csv("./kegg_df.csv")
target_df <- read.csv("./target_df.csv")

result <- generate_target_types(kegg_df, target_df)
node_df <- result[[1]]
type_df <- result[[2]]
type_counts <- result[[4]]
target_num <- type_counts["target"]
compound_num <- type_counts["compound"] 
}

}
