% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-bioranger.Rmd.
\name{import_biorange}
\alias{import_biorange}
\title{导入Python biorange模块}
\usage{
import_biorange()
}
\value{
导入的biorange模块
}
\description{
此函数尝试使用reticulate包导入Python的biorange模块。
}
\details{
该函数使用reticulate包的import函数来加载Python的biorange模块。
如果导入失败，函数将抛出一个错误，提示用户确保正确安装了biorange模块。
}
\examples{
\dontrun{
biorange <- import_biorange()
}
}