% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_environment.R
\name{initialize_environment}
\alias{initialize_environment}
\title{初始化包的Python环境}
\usage{
initialize_environment(venv_name = "r-reticulate", min_python_version = "3.10", required_packages = c("biorange"))
}
\arguments{
\item{venv_name}{将要创建或使用的虚拟环境名称。}

\item{min_python_version}{所需的最低Python版本。}

\item{required_packages}{需要安装的Python包的字符向量。}
}
\value{
返回一个指示环境设置状态的消息。
}
\description{
该函数设置包所需的Python环境。它会在必要时安装Miniconda，创建虚拟环境，并安装所需的Python包。
}
\details{
该函数检查Python是否可用，如果不可用则安装Miniconda；如果虚拟环境不存在则创建虚拟环境，并安装所需的Python包。
}
\examples{
\dontrun{
initialize_environment()
}
}
\keyword{Python,}
\keyword{reticulate}
\keyword{环境,}
\keyword{设置,}
