% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{node_data_339}
\alias{node_data_339}
\title{节点关系数据集（339行）}
\format{
一个包含339行和2列的数据框：
\describe{
\item{node1}{字符型，表示关系中的第一个节点。}
\item{node2}{字符型，表示关系中的第二个节点。}
}
}
\source{
\link{数据来源，如果有的话}
}
\usage{
data(node_data_339)
}
\description{
该数据集包含节点之间的关系信息，共339行，两列：\code{node1}和\code{node2}。
}
\examples{
data(node_data_339)
head(node_data_339)
}
\keyword{datasets}
