% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_core_targets.R
\name{ppi_core_targets}
\alias{ppi_core_targets}
\title{PPI核心靶点网络图}
\usage{
ppi_core_targets(
  nodes,
  degree_df,
  output_dir = "./results/output/pp6",
  file_name = "PPI_network_degree_top",
  top_nodes = 10
)
}
\arguments{
\item{nodes}{Data frame. Contains node information for the PPI network.}

\item{degree_df}{Data frame. Contains degree centrality measures for nodes.}

\item{degree_target_num}{Numeric. The target number of top nodes to select based on degree.}

\item{core_output_dir}{Character. Directory path where output files will be saved.
Default is "./results/output/pp6".}

\item{file_name_prefix}{Character. Prefix for output file names.
Default is "PPI_network_degree_top".}
}
\value{
Results from generate_core_targets function, typically including information
about the core targets identified.
}
\description{
这个函数将PPI结果按照degree排序筛选靶点，绘制网络图
}
\details{
Identify and Generate Core Targets from PPI Network

This function identifies top nodes based on degree centrality in a protein-protein
interaction (PPI) network and generates core targets from these nodes.
}
\examples{
\dontrun{
kegg_top_nodes <- target_top_nodes(10) # Get top 10 nodes
my_nodes <- nodes_df_88
my_degree_df <- degree_df_88
results <- generate_core_targets(
    nodes = my_nodes,
    degree_df = my_degree_df,
    output_dir = "./my_output",
    file_name = "PPI_network_degree_top10",
    top_nodes = 100
)
}

}
