% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_stitch_inchikey_targets.R
\name{predict_stitch_inchikey_targets}
\alias{predict_stitch_inchikey_targets}
\title{使用InChIKeys预测STITCH靶点}
\usage{
predict_stitch_inchikey_targets(inchikeys)
}
\arguments{
\item{inchikeys}{一个包含InChIKeys的字符向量}
}
\value{
一个包含预测STITCH靶点的数据框
}
\description{
此函数对一批InChIKeys进行STITCH靶点预测。
}
\details{
该函数使用biorange模块的stich_inchikey_target函数对InChIKeys进行靶点预测。
结果会被返回为一个数据框。
}
\examples{
\dontrun{
inchikeys <- c("BSYNRYMUTXBXSQ-UHFFFAOYSA-N", "LFQSCWFLJHTTHZ-UHFFFAOYSA-N")
stitch_inchikey_targets <- predict_stitch_inchikey_targets(inchikeys)
print(stitch_inchikey_targets)
}
}
