% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-bioranger.Rmd.
\name{predict_tcmsp_batch_targets}
\alias{predict_tcmsp_batch_targets}
\title{批量预测TCMSP靶点}
\usage{
predict_tcmsp_batch_targets(smiles)
}
\arguments{
\item{smiles}{一个包含SMILES字符串的字符向量}
}
\value{
一个包含预测TCMSP靶点的数据框
}
\description{
此函数对一批SMILES字符串进行TCMSP靶点预测。
}
\details{
该函数使用biorange模块的tcmsp_smiles_target函数对每个输入的SMILES字符串进行靶点预测。
结果会被合并成一个数据框返回。
}
\examples{
\dontrun{
smiles <- c("CC(=O)OC1=CC=CC=C1C(=O)O", "CCO")
tcmsp_targets <- predict_tcmsp_batch_targets(smiles)
print(tcmsp_targets)
}
}