% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_tcmsp_inchikey_targets.R
\name{predict_tcmsp_inchikey_targets}
\alias{predict_tcmsp_inchikey_targets}
\title{使用InChIKeys预测TCMSP靶点}
\usage{
predict_tcmsp_inchikey_targets(inchikeys)
}
\arguments{
\item{inchikeys}{一个包含InChIKeys的字符向量}
}
\value{
一个包含预测TCMSP靶点的数据框
}
\description{
此函数对一批InChIKeys进行TCMSP靶点预测。
}
\details{
该函数使用biorange模块的tcmsp_inchikey_target函数对输入的InChIKeys进行靶点预测。
结果会被返回为一个数据框。
}
\examples{
\dontrun{
inchikeys <- c("BSYNRYMUTXBXSQ-UHFFFAOYSA-N", "LFQSCWFLJHTTHZ-UHFFFAOYSA-N")
tcmsp_inchikey_targets <- predict_tcmsp_inchikey_targets(inchikeys)
print(tcmsp_inchikey_targets)
}
}
