% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-bioranger.Rmd.
\name{predict_ttd_disease_targets}
\alias{predict_ttd_disease_targets}
\title{预测TTD疾病靶点}
\usage{
predict_ttd_disease_targets(diseases)
}
\arguments{
\item{diseases}{一个表示疾病的字符串}
}
\value{
一个包含预测疾病靶点的数据框
}
\description{
此函数使用TTD（治疗靶点数据库）预测疾病靶点。
}
\details{
该函数使用biorange模块的ttd_disease_target函数对输入的疾病进行靶点预测。
结果会被返回为一个数据框。
}
\examples{
\dontrun{
ttd_targets <- predict_ttd_disease_targets("癌症")
print(ttd_targets)
}
}