% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-bioranger.Rmd.
\name{run_ppi_analysis}
\alias{run_ppi_analysis}
\title{使用string构建PPI网络}
\usage{
run_ppi_analysis(gene_names_file, output_dir)
}
\arguments{
\item{gene_names_file}{A character string specifying the path to the CSV file containing gene names.}

\item{output_dir}{A character string specifying the directory where the output will be saved.}
}
\description{
This function performs protein-protein interaction (PPI) analysis using the BioRange package.
It imports the necessary module and calls the appropriate function with the specified gene names
file and output directory.
}