% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppi_total_network.R
\name{total_network_plot}
\alias{total_network_plot}
\title{ppi_total_network}
\usage{
total_network_plot(kegg_data, metabolites_list, degree_table, disease, formula,output_dir)
}
\arguments{
\item{kegg_data}{包含KEGG通路数据的数据框。}

\item{metabolites_list}{包含代谢物及其靶标列表的数据框。}

\item{degree_table}{包含靶点degree值的数据框。}

\item{disease}{指定感兴趣的疾病的字符串，例如“Lung cancer”。}

\item{formula}{指定方剂的首字母缩写字符串，例如四君子汤对应“SJZT”。}
\item{output_dir}{结果保存地址。}
}
\value{
2个包含网络图分析生成的输出文件的列表，total_node_file和total_type_file。
}
\description{
该函数根据KEGG数据、代谢物列表和节点度信息表，创建汇总网络图（total_naework）的节点和类型文件并生成网络图。
}
\details{
该函数处理输入数据以生成综合网络图，帮助可视化交互关系。
}
\examples{
kegg_res <- read.csv("/path/to/kegg_res.csv")
metabolites_list <- read.csv("/path/to/metabolite_targets.csv")
degree_table <- read.csv("/path/to/string_node_degree.csv")
disease_name <- "Lung cancer"
formula_name <- "SJZT"
output_dir <-"/path/ppi/total_network_pathway"
output_files <- total_network_plot(kegg_data, metabolites_list, degree_table, disease, formula,output_dir)
print(output_files)
}
\keyword{KEGG,}
\keyword{代谢物,}
\keyword{度}
\keyword{绘图,}
\keyword{网络,}
