% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{type_data_146}
\alias{type_data_146}
\title{节点类型数据集}
\format{
一个包含146行和2列的数据框：
\describe{
\item{node}{字符型，表示节点的名称。}
\item{type}{字符型，表示节点的类型。}
}
}
\source{
\link{数据来源，如果有的话}
}
\usage{
data(type_data_146)
}
\description{
该数据集包含节点的类型信息，两列：\code{node}和\code{type}。
}
\examples{
data(type_data_146)
head(type_data_146)
}
\keyword{datasets}
