% Generated by roxygen2: do not edit by hand
% Please edit documentation in create-bioranger.Rmd.
\name{predict_genecards_disease_targets}
\alias{predict_genecards_disease_targets}
\title{预测GeneCards疾病靶点}
\usage{
predict_genecards_disease_targets(query_string)
}
\arguments{
\item{query_string}{一个表示疾病查询的字符串}
}
\value{
一个包含预测疾病靶点的数据框
}
\description{
此函数使用GeneCards数据库预测疾病靶点。
}
\details{
该函数使用biorange模块的genecards_disease_target函数对输入的疾病查询字符串进行靶点预测。
结果会被返回为一个数据框。注意：使用此功能需要自行下载GeneCards数据。
}
\examples{
\dontrun{
genecards_targets <- predict_genecards_disease_targets("癌症")
print(genecards_targets)
}
}