% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ppi_analysis.R
\name{run_ppi_analysis}
\alias{run_ppi_analysis}
\title{使用string构建PPI网络}
\usage{
run_ppi_analysis(gene_names_file, output_dir)
}
\arguments{
\item{gene_names_file}{A chr specifying the path to the CSV file}

\item{output_dir}{A character string specifying the directory}
}
\description{
This function performs protein-protein interaction (PPI) analysis
using the BioRange package.It imports the necessary module and calls the
appropriate function with the specified gene names file and output directory.
}
